adjust user service bootstrapping
