+++ apps/user-service/src/bootstrap.ts
@@ -1,5 +1,8 @@
+import { ensureUsersProfileAttributes } from '@social-media-platform/keycloak-utils';
 import { migrator, sequelize } from './config';
 import Friendship from './models/friendship';
+import { ensureDefaultPrivacySettings } from './utils';
+import { attributes, groups, defaultUsersIds } from './constants';
 
 export default async function bootstrap() {
   await sequelize.authenticate();
@@ -9,4 +12,8 @@ export default async function bootstrap() {
   console.log('Database up to date!');
 
   Friendship.initializeHooks();
+
+  await ensureUsersProfileAttributes(attributes, groups);
+
+  await Promise.all(defaultUsersIds.map(ensureDefaultPrivacySettings));
 }
